refactor: components를 index.js에서 export하도록 변경
