+++ src/components/orderListItem/OrderListItem.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import Button, { BUTTON_TYPE } from '../button/Button';
-import ProductImage, { PRODUCT_IMAGE_TYPE } from '../productImage/ProductImage';
-import { COLOR } from '../../constants/color';
 import PropTypes from 'prop-types';
+import { Button, BUTTON_TYPE } from '../';
+import { ProductImage, PRODUCT_IMAGE_TYPE } from '../';
+import { COLOR } from '../../constants/color';
 
 export const ORDER_LIST_ITEM_TYPE = Object.freeze({
   ORDER_PAYMENT: 'ORDER_PAYMENT',
