+++ src/index.js
@@ -1,14 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import { applyMiddleware, createStore } from 'redux';
-import { Provider } from 'react-redux';
-import thunk from 'redux-thunk';
-import { composeWithDevTools } from 'redux-devtools-extension';
 import App from './App';
-import rootReducer from './modules';
-
-const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
+import store from './modules/store';
+import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
 
 ReactDOM.render(
   <React.StrictMode>
