+++ src/pages/ProductList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import ProductListItem from '../components/productListItem/ProductListItem';
+import { useSelector } from 'react-redux';
+import { ProductListItem } from '../components';
 
 const Container = styled.ul`
   display: grid;
